src/requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.21.0
MetaTrader5>=5.0.45

# Web Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.4.0
dash-table>=5.0.0
plotly>=5.15.0

# AI and Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# Logging and Monitoring
colorama>=0.4.6

# Data Analysis and Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Technical Analysis
TA-Lib>=0.4.25

# Configuration and Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Production Deployment
gunicorn>=21.2.0

# Additional Dependencies
requests>=2.31.0
python-dateutil>=2.8.2
